refactor(phase3c): consolidate provider base classes into ai/core/BaseProvider.ts
- new ai/core/BaseProvider.ts is the single home for the base-class family
- extract shared root AbstractProviderCore (model + ctor + convertMcpTools<T>),
  removing the duplicate convertMcpTools<T> declaration across the two old bases
- AbstractBaseProvider (sendChatMessage family) and BaseAIProvider/BaseOpenAIResponseProvider
  (completions family) now both extend the shared root; distinct surfaces kept separate (ISP/LSP)
- old api/baseProvider.ts and providers/BaseAIProvider.ts become re-export shims (downstream imports unchanged)

Zero behavior change: pure relocation + shared-root extraction. type-check/test(32)/build green.
